Change location of date validator
