+++ webapp/app/forms/validators.py
@@ -1,5 +1,4 @@
 from collections import Counter
-import datetime
 
 
 from schwifty import IBAN
@@ -117,74 +116,3 @@ class ValidUnlockCodeCharacterSet:
         for char in input_str:
             if char not in VALID_UNLOCK_CODE_CHARACTERS:
                 raise ValidationError(_('validate.invalid-character'))
-
-class ValidDateOf:
-    # Validate a date between 01.01.1900 and today
-    # Possible validation errors:
-    # - Date is invalid (99.99.2021)
-    # - Date is incomplete (__.__.2021)
-    # - Date is in the future
-    # - Date is is before 01.01.1990
-    
-    def __init__(self, message_incomplete, message_incorrect, message_in_the_future, message_to_far_in_past) -> None:
-        self.message_incomplete = message_incomplete
-        self.message_incorrect = message_incorrect
-        self.message_in_the_future = message_in_the_future
-        self.message_to_far_in_past = message_to_far_in_past
-        
-
-    def __call__(self, form, field):
-        raw_input_date = field._value()
-        elster_min_date = datetime.date(1900, 1, 1)
-        
-        day, month, year = None, None, None
-
-        if isinstance(raw_input_date, str):
-            day, month, year = raw_input_date.split('.')
-        elif isinstance(raw_input_date, list):
-            if len(raw_input_date) != 3:
-                raise ValidationError(self.message_incomplete)
-
-            day, month, year = raw_input_date
-
-        if not day or not month or not year: 
-            raise ValidationError(self.message_incomplete)
-            
-        try:
-            input_date = datetime.date(int(year), int(month), int(day))
-        except ValueError:
-            raise ValidationError(self.message_incorrect)
-
-        if input_date > datetime.date.today():
-            raise ValidationError(self.message_in_the_future)
-
-        if input_date < elster_min_date:
-            raise ValidationError(self.message_to_far_in_past)
-
-class ValidDateOfBirth(ValidDateOf):
-    def __init__(self) -> None:
-        super().__init__(message_incomplete = _('validate.date-of-birth-incomplete'), 
-            message_incorrect = _('validate.date-of-incorrect'), 
-            message_in_the_future = _('validate.date-of-in-the-future'), 
-            message_to_far_in_past = _('validate.date-of-to-far-in-past'))
-class ValidDateOfDeath(ValidDateOf):
-    def __init__(self) -> None:
-        super().__init__(message_incomplete = _('validate.date-of-death-incomplete'), 
-            message_incorrect = _('validate.date-of-incorrect'), 
-            message_in_the_future = _('validate.date-of-in-the-future'), 
-            message_to_far_in_past = _('validate.date-of-to-far-in-past'))
-        
-class ValidDateOfMarriage(ValidDateOf):
-    def __init__(self) -> None:
-        super().__init__(message_incomplete = _('validate.date-of-marriage-incomplete'), 
-            message_incorrect = _('validate.date-of-incorrect'), 
-            message_in_the_future = _('validate.date-of-in-the-future'), 
-            message_to_far_in_past = _('validate.date-of-to-far-in-past'))
-        
-
-class ValidDateOfDivorce(ValidDateOf):
-    def __init__(self) -> None:
-        super().__init__(message_incomplete = _('validate.date-of-incomplete'), 
-            message_incorrect = _('validate.date-of-incorrect'), 
-            message_in_the_future = _('validate.date-of-in-the-future'), 
-            message_to_far_in_past = _('validate.date-of-to-far-in-past'))
