+++ webapp/tests/forms/test_validators.py
@@ -3,8 +3,8 @@ import unittest
 from wtforms import IntegerField, ValidationError, StringField
 
 from app.forms import SteuerlotseBaseForm
-from app.forms.fields import SteuerlotseDateField, UnlockCodeField
-from app.forms.validators import IntegerLength, ValidDateOfBirth, ValidIdNr, DecimalOnly, ValidElsterCharacterSet, ValidUnlockCode, \
+from app.forms.fields import UnlockCodeField
+from app.forms.validators import IntegerLength, ValidIdNr, DecimalOnly, ValidElsterCharacterSet, ValidUnlockCode, \
     ValidUnlockCodeCharacterSet
